feat: 인연 충전(POST /api/points) 구현
- MemberPoint·PointTransaction 애그리거트, PointQuantity VO 추가
- PaymentApprover 포트 + Router 구조 도입(Mock 1차), PaymentMethod enum 정의
- DB 비관적 잠금(row lock)과 멱등키 UNIQUE로 동시성·중복 방지
- expectedPrice 서버 재계산 비교로 가격 변경 탐지
- POST /api/points만 authenticated, GET은 기존 permitAll 유지
- REST Docs: 201/402/404/409 및 유효성 케이스 문서화

Plan: docs/plan/202604/point-charge.plan.md

+++ boot/ma-boot-web/src/main/kotlin/com/konkuk/ma/domain/point/api/response/PointProductResponse.kt
@@ -17,7 +17,7 @@ class PointProductResponse(
             return PointProductResponse(
                 pointProductId = product.pointProduct.pointProductId,
                 name = product.pointProduct.name,
-                quantity = product.pointProduct.quantity,
+                quantity = product.pointProduct.quantity.toInt(),
                 price = product.pointProduct.price.toInt(),
                 discountedPrice = product.discountedPrice().toInt(),
                 discountRate = product.discountRate().toInt(),
